some bug fix

+++ lib/libft/ft_putnbr_fd.c
@@ -6,17 +6,17 @@
 /*   By: lrio <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 17:13:17 by lrio              #+#    #+#             */
-/*   Updated: 2024/01/02 17:20:45 by lrio             ###   ########.fr       */
+/*   Updated: 2024/01/03 16:25:45 by lrio             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putnbr_fd(int n, int fd)
+int	ft_putnbr_fd(long n, int fd)
 {
-	long	num;
+	long long	num;
 
-	num = (long)n;
+	num = (long long)n;
 	if (num < 0)
 	{
 		if (-1 == ft_putchar_fd('-', fd))
